Change forge descriptor to use new gcli_id type instead of plain int
The reasoning behind this is that if we decide to change the ID
types or if we realise that it is wrong how we handle it at the
moment we just have one typedef that affects all uses. Especially
for GitLab this seems to be important as the IDs returned by it
should be ulongs (learned from experience).

+++ include/gcli/gcli.h
@@ -45,7 +45,7 @@ typedef enum gcli_forge_type {
 	GCLI_FORGE_GITEA,
 } gcli_forge_type;
 
-typedef unsigned int gcli_id;
+typedef unsigned long gcli_id;
 
 #ifdef IN_LIBGCLI
 #include <gcli/ctx.h>
